REFACTOR helper functions
