--- src/__index.js
@@ -1,532 +1,0 @@
-import Phaser from 'phaser';
-
-var Message = new Phaser.Class({
-
-  Extends: Phaser.GameObjects.Container,
-
-  initialize:
-    function Message(scene, events) {
-      Phaser.GameObjects.Container.call(this, scene, 160, 30);
-      var graphics = this.scene.add.graphics();
-      this.add(graphics);
-      graphics.lineStyle(1, 0xffffff, 0.8);
-      graphics.fillStyle(0x031f4c, 0.3);
-      graphics.strokeRect(-90, -15, 180, 30);
-      graphics.fillRect(-90, -15, 180, 30);
-      this.text = new Phaser.GameObjects.Text(scene, 0, 0, "", { color: '#ffffff', align: 'center', fontSize: 13, wordWrap: { width: 160, useAdvancedWrap: true } });
-      this.add(this.text);
-      this.text.setOrigin(0.5);
-      events.on("Message", this.showMessage, this);
-      this.visible = false;
-    },
-  showMessage: function (text) {
-    this.text.setText(text);
-    this.visible = true;
-    if (this.hideEvent)
-      this.hideEvent.remove(false);
-    this.hideEvent = this.scene.time.addEvent({ delay: 2000, callback: this.hideMessage, callbackScope: this });
-  },
-  hideMessage: function () {
-    this.hideEvent = null;
-    this.visible = false;
-  }
-});
-
-let Unit = new Phaser.Class({
-  Extends: Phaser.GameObjects.Sprite,
-
-  initialize:
-
-    function Unit(scene, x, y, texture, frame, type, hp, damage) {
-      Phaser.GameObjects.Sprite.call(this, scene, x, y, texture, frame)
-      this.type = type;
-      this.maxHp = this.hp = hp;
-      this.damage = damage; // default damage                
-    },
-  attack: function (target) {
-    target.takeDamage(this.damage);
-  },
-  takeDamage: function (damage) {
-    this.hp -= damage;
-  }
-});
-
-let Enemy = new Phaser.Class({
-  Extends: Unit,
-
-  initialize:
-    function Enemy(scene, x, y, texture, frame, type, hp, damage) {
-      Unit.call(this, scene, x, y, texture, frame, type, hp, damage);
-    }
-});
-
-let PlayerCharacter = new Phaser.Class({
-  Extends: Unit,
-
-  initialize:
-    function PlayerCharacter(scene, x, y, texture, frame, type, hp, damage) {
-      Unit.call(this, scene, x, y, texture, frame, type, hp, damage);
-      // flip the image so I don't have to edit it manually
-      this.flipX = true;
-
-      this.setScale(2);
-    }
-});
-
-// menu 
-let MenuItem = new Phaser.Class({
-  Extends: Phaser.GameObjects.Text,
-
-  initialize:
-
-    function MenuItem(x, y, text, scene) {
-      Phaser.GameObjects.Text.call(this, scene, x, y, text, { color: '#ffffff', align: 'left', fontSize: 15 });
-    },
-
-  select: function () {
-    this.setColor('#f8ff38');
-  },
-
-  deselect: function () {
-    this.setColor('#ffffff');
-  }
-})
-
-let Menu = new Phaser.Class({
-  Extends: Phaser.GameObjects.Container,
-
-  initialize:
-
-    function Menu(x, y, scene, heroes) {
-      Phaser.GameObjects.Container.call(this, scene, x, y);
-      this.menuItems = [];
-      this.menuItemIndex = 0;
-      this.heroes = heroes;
-      this.x = x;
-      this.y = y;
-    },
-  addMenuItem: function (unit) {
-    var menuItem = new MenuItem(0, this.menuItems.length * 20, unit, this.scene);
-    this.menuItems.push(menuItem);
-    this.add(menuItem);
-  },
-  moveSelectionUp: function () {
-    this.menuItems[this.menuItemIndex].deselect();
-    this.menuItemIndex--;
-    if (this.menuItemIndex < 0)
-      this.menuItemIndex = this.menuItems.length - 1;
-    this.menuItems[this.menuItemIndex].select();
-  },
-  moveSelectionDown: function () {
-    this.menuItems[this.menuItemIndex].deselect();
-    this.menuItemIndex++;
-    if (this.menuItemIndex >= this.menuItems.length)
-      this.menuItemIndex = 0;
-    this.menuItems[this.menuItemIndex].select();
-  },
-  // select the menu as a whole and an element with index from it
-  select: function (index) {
-    if (!index)
-      index = 0;
-    this.menuItems[this.menuItemIndex].deselect();
-    this.menuItemIndex = index;
-    this.menuItems[this.menuItemIndex].select();
-  },
-  // deselect this menu
-  deselect: function () {
-    this.menuItems[this.menuItemIndex].deselect();
-    this.menuItemIndex = 0;
-  },
-  confirm: function () {
-    // wen the player confirms his slection, do the action
-  },
-  clear: function () {
-    for (var i = 0; i < this.menuItems.length; i++) {
-      this.menuItems[i].destroy();
-    }
-    this.menuItems.length = 0;
-    this.menuItemIndex = 0;
-  },
-  remap: function (units) {
-    this.clear();
-    for (var i = 0; i < units.length; i++) {
-      var unit = units[i];
-      this.addMenuItem(unit.type);
-    }
-  }
-});
-
-let HeroesMenu = new Phaser.Class({
-  Extends: Menu,
-
-  initialize:
-
-    function HeroesMenu(x, y, scene) {
-      Menu.call(this, x, y, scene);
-    }
-});
-
-let ActionsMenu = new Phaser.Class({
-  Extends: Menu,
-
-  initialize:
-
-    function ActionsMenu(x, y, scene) {
-      Menu.call(this, x, y, scene);
-      this.addMenuItem('Attack');
-    },
-  confirm: function () {
-    this.scene.events.emit('SelectEnemies');
-  }
-
-});
-
-let EnemiesMenu = new Phaser.Class({
-  Extends: Menu,
-
-  initialize:
-
-    function EnemiesMenu(x, y, scene) {
-      Menu.call(this, x, y, scene);
-    },
-  confirm: function () {
-    this.scene.events.emit("Enemy", this.menuItemIndex);
-  }
-});
-
-var BootScene = new Phaser.Class({
-
-  Extends: Phaser.Scene,
-
-  initialize:
-
-    function BootScene() {
-      Phaser.Scene.call(this, { key: 'BootScene' });
-    },
-
-  preload() {
-    // load the resources here
-    // map tiles
-    //this.load.image('tiles', 'src/assets/map/spritesheet.png');
-
-    // map in json format
-    //this.load.tilemapTiledJSON('map', 'src/assets/map/map.json');
-
-    // our two characters
-    //this.load.spritesheet('player', 'src/assets/RPG_assets.png', { frameWidth: 16, frameHeight: 16 });
-
-    // load resources
-    this.load.spritesheet('player', 'src/assets/RPG_assets.png', { frameWidth: 16, frameHeight: 16 });
-    this.load.image('dragonblue', 'src/assets/dragonblue.png');
-    this.load.image('dragonorrange', 'src/assets/dragonorrange.png');
-
-  },
-
-  create() {
-    //this.scene.start('BattleScene')
-    this.scene.start('WorldScene');
-  }
-});
-
-const BattleScene = new Phaser.Class({
-
-  Extends: Phaser.Scene,
-
-  initialize:
-
-    function BattleScene() {
-      Phaser.Scene.call(this, { key: 'BattleScene' });
-    },
-  create() {
-    // change the background to green
-    this.cameras.main.setBackgroundColor('rgba(0, 200, 0, 0.5)');
-
-    // player character - warrior
-    var warrior = new PlayerCharacter(this, 250, 50, 'player', 1, 'Warrior', 100, 20);
-    this.add.existing(warrior);
-
-    // player character - mage
-    let mage = new PlayerCharacter(this, 250, 100, 'player', 4, 'Mage', 80, 8);
-    this.add.existing(mage);
-
-    let dragonblue = new Enemy(this, 50, 50, 'dragonblue', null, 'Dragon', 50, 3);
-    this.add.existing(dragonblue);
-
-    let dragonOrange = new Enemy(this, 50, 100, 'dragonorrange', null, 'Dragon2', 50, 3);
-    this.add.existing(dragonOrange);
-
-    // array with heroes
-    this.heroes = [warrior, mage];
-    // array with enemies
-    this.enemies = [dragonblue, dragonOrange];
-    // array with both parties, who will attack
-    this.units = this.heroes.concat(this.enemies);
-
-    // Run UI Scene at the same time
-    this.scene.launch('UIScene');
-
-    this.index = -1
-  },
-  receivePlayerSelection: function (action, target) {
-    if (action == 'attack') {
-      this.units[this.index].attack(this.enemies[target]);
-    }
-    this.time.addEvent({ delay: 3000, callback: this.nextTurn, callbackScope: this });
-  },
-  nextTurn: function () {
-    this.index++;
-    // if there are no more units, we start again from the first one
-    if (this.index >= this.units.length) {
-      this.index = 0;
-    }
-    if (this.units[this.index]) {
-      // if its player hero
-      if (this.units[this.index] instanceof PlayerCharacter) {
-        this.events.emit('PlayerSelect', this.index);
-      } else { // else if its enemy unit
-        // pick random hero
-        var r = Math.floor(Math.random() * this.heroes.length);
-        // call the enemy's attack function 
-        this.units[this.index].attack(this.heroes[r]);
-        // add timer for the next turn, so will have smooth gameplay
-        this.time.addEvent({ delay: 3000, callback: this.nextTurn, callbackScope: this });
-      }
-    }
-  },
-});
-
-var UIScene = new Phaser.Class({
-
-  Extends: Phaser.Scene,
-
-  initialize:
-
-    function UIScene() {
-      Phaser.Scene.call(this, { key: 'UIScene' });
-    },
-
-  create() {
-    this.graphics = this.add.graphics();
-    this.graphics.lineStyle(1, 0xffffff);
-    this.graphics.fillStyle(0x031f4c, 1);
-    this.graphics.strokeRect(2, 150, 90, 100);
-    this.graphics.fillRect(2, 150, 90, 100);
-    this.graphics.strokeRect(95, 150, 90, 100);
-    this.graphics.fillRect(95, 150, 90, 100);
-    this.graphics.strokeRect(188, 150, 130, 100);
-    this.graphics.fillRect(188, 150, 130, 100);
-
-    // basic container to hold all menus
-    this.menus = this.add.container();
-
-    this.heroesMenu = new HeroesMenu(195, 153, this);
-    this.actionsMenu = new ActionsMenu(100, 153, this);
-    this.enemiesMenu = new EnemiesMenu(8, 153, this);
-
-    // the currently selected menu 
-    this.currentMenu = this.actionsMenu;
-
-    // add menus to the container
-    this.menus.add(this.heroesMenu);
-    this.menus.add(this.actionsMenu);
-    this.menus.add(this.enemiesMenu);
-
-    this.battleScene = this.scene.get('BattleScene')
-
-    this.remapHeroes()
-    this.remapEnemies()
-
-    this.input.keyboard.on('keydown', this.onKeyInput, this)
-    this.battleScene.events.on('PlayerSelect', this.onPlayerSelect, this)
-
-    this.events.on('SelectEnemies', this.onSelectEnemies, this)
-    this.events.on("Enemy", this.onEnemy, this)
-    this.battleScene.nextTurn()
-    this.message = new Message(this, this.battleScene.events);
-    this.add.existing(this.message);
-  },
-  onEnemy: function (index) {
-    this.heroesMenu.deselect();
-    this.actionsMenu.deselect();
-    this.enemiesMenu.deselect();
-    this.currentMenu = null;
-    this.battleScene.receivePlayerSelection('attack', index);
-  },
-  onSelectEnemies: function () {
-    this.currentMenu = this.enemiesMenu;
-    this.enemiesMenu.select(0);
-  },
-  remapHeroes: function () {
-    var heroes = this.battleScene.heroes;
-    this.heroesMenu.remap(heroes);
-  },
-  remapEnemies: function () {
-    var enemies = this.battleScene.enemies;
-    this.enemiesMenu.remap(enemies);
-  },
-  onKeyInput: function (event) {
-    if (this.currentMenu) {
-      if (event.code === "ArrowUp") {
-        this.currentMenu.moveSelectionUp();
-      } else if (event.code === "ArrowDown") {
-        this.currentMenu.moveSelectionDown();
-      } else if (event.code === "ArrowRight" || event.code === "Shift") {
-
-      } else if (event.code === "Space" || event.code === "ArrowLeft") {
-        this.currentMenu.confirm();
-      }
-    }
-  },
-  onPlayerSelect: function (id) {
-    this.heroesMenu.select(id);
-    this.actionsMenu.select(0);
-    this.currentMenu = this.actionsMenu;
-  },
-});
-
-var WorldScene = new Phaser.Class({
-
-  Extends: Phaser.Scene,
-
-  initialize:
-
-    function WorldScene() {
-      Phaser.Scene.call(this, { key: 'WorldScene' });
-    },
-  preload() {
-
-  },
-  create() {
-    // create your world and bounds here
-    var map = this.make.tilemap({ key: 'map' });
-    var tiles = map.addTilesetImage('spritesheet', 'tiles');
-
-    var grass = map.createStaticLayer('Grass', tiles, 0, 0);
-    var obstacles = map.createStaticLayer('Obstacles', tiles, 0, 0);
-    obstacles.setCollisionByExclusion([-1]);
-
-    // set player sprite with bounds
-    this.player = this.physics.add.sprite(50, 100, 'player', 6)
-
-    this.physics.world.bounds.width = map.widthInPixels
-    this.physics.world.bounds.height = map.heightInPixels
-    this.player.setCollideWorldBounds(true)
-
-    // enable keys
-    this.cursors = this.input.keyboard.createCursorKeys()
-
-    // camera setup
-    this.cameras.main.setBounds(0, 0, map.widthInPixels, map.heightInPixels)
-    this.cameras.main.startFollow(this.player)
-    this.cameras.main.roundPixels = true
-
-    // player animations
-    //  animation with key 'left', we don't need left and right as we will use one and flip the sprite
-    this.anims.create({
-      key: 'left',
-      frames: this.anims.generateFrameNumbers('player', { frames: [1, 7, 1, 13] }),
-      frameRate: 10,
-      repeat: -1
-    });
-
-    // animation with key 'right'
-    this.anims.create({
-      key: 'right',
-      frames: this.anims.generateFrameNumbers('player', { frames: [1, 7, 1, 13] }),
-      frameRate: 10,
-      repeat: -1
-    });
-    this.anims.create({
-      key: 'up',
-      frames: this.anims.generateFrameNumbers('player', { frames: [2, 8, 2, 14] }),
-      frameRate: 10,
-      repeat: -1
-    });
-    this.anims.create({
-      key: 'down',
-      frames: this.anims.generateFrameNumbers('player', { frames: [0, 6, 0, 12] }),
-      frameRate: 10,
-      repeat: -1
-    });
-
-    // player can bump into trees 
-    this.physics.add.collider(this.player, obstacles);
-
-    // dummy game objects
-    this.spawns = this.physics.add.group({ classType: Phaser.GameObjects.Zone });
-    for (var i = 0; i < 30; i++) {
-      var x = Phaser.Math.RND.between(0, this.physics.world.bounds.width);
-      var y = Phaser.Math.RND.between(0, this.physics.world.bounds.height);
-      // parameters are x, y, width, height
-      this.spawns.create(x, y, 20, 20);
-    }
-    this.physics.add.overlap(this.player, this.spawns, this.onMeetEnemy, false, this);
-
-    onMeetEnemy: (player, zone) => {
-
-      zone.x = Phaser.Math.RND.between(0, this.physics.world.bounds.width)
-      zone.y = Phaser.Math.RND.between(0, this.physics.world.bounds.height)
-
-      // shake the world
-      this.cameras.main.shake(300)
-
-      // start battle
-      createHtmlTagObject.scene.switch('BattleScene')
-
-    }
-  },
-
-  update(time, delta) {
-    this.player.body.setVelocity(0)
-
-
-    // movement in space
-    // horizontal movement
-    if (this.cursors.left.isDown) {
-      this.player.body.setVelocityX(-80)
-    } else if (this.cursors.right.isDown) {
-      this.player.body.setVelocityX(80)
-    }
-    // vertical movement
-    if (this.cursors.up.isDown) {
-      this.player.body.setVelocityY(-80)
-    } else if (this.cursors.down.isDown) {
-      this.player.body.setVelocityY(80)
-    }
-
-    // local sprite motion
-    // player movement animations 
-    if (this.cursors.left.isDown) {
-      this.player.anims.play('left', true);
-    }
-    else if (this.cursors.right.isDown) {
-      this.player.anims.play('right', true);
-    }
-    else if (this.cursors.up.isDown) {
-      this.player.anims.play('up', true);
-    }
-    else if (this.cursors.down.isDown) {
-      this.player.anims.play('down', true);
-    }
-    else {
-      this.player.anims.stop();
-    }
-  }
-});
-
-const config = {
-  type: Phaser.AUTO,
-  parent: 'content',
-  width: 320,
-  height: 240,
-  zoom: 2,
-  pixelArt: true,
-  physics: {
-    default: 'arcade',
-    arcade: {
-      gravity: { y: 0 }
-    }
-  },
-  scene: [BootScene, WorldScene, BattleScene, UIScene]
-};
-var game = new Phaser.Game(config);
