+++ src/scenes/RankingScene.js
@@ -1,18 +1,7 @@
 import Phaser from 'phaser';
 import Button from '../objects/Button';
 import config from '../config/config';
-
-
-const getRanking = async () => {
-  let apiURL = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/tB4vCeH1M2zdNORMm4cr/scores'
-  try {
-    let res = await fetch(apiURL)
-    res = await res.json()
-    return res
-  } catch {
-    throw new Error('API score fetch failed')
-  }
-}
+import getRanking from '../helpers/ranking';
